pyproject.toml
[project]
name = "genesis-medai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "moviepy>=2.2.1",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
